Add updates
Return a char from a command to update it. This is not as good as I'd
like it to be, but it'll do.
I'm considering returnign a class like `UpdateChar(newchar)` or
something to better signify intent.

Also, load perks differently now that YAML::XS serializes them.

Add commands to reclass/level a character.